project_name/prod/requirements.txt
mlflow==2.9.2
flatdict==4.0.1
fastapi==0.109.2
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-dotenv